+++ lib/csvloader.dart
@@ -1,5 +1,5 @@
 library csvloader;
 
-export 'src/csv_data.dart' show CsvData;
+export 'src/csv_data.dart' hide CsvDataImpl;
 export 'src/csv_loader.dart';
 export 'src/exceptions.dart';