+++ lib/src/_csv_processor.dart
@@ -1,22 +1,23 @@
 import 'dart:async';
 
-import 'package:csvloader/src/csv_headers.dart';
+import 'package:csvloader/src/_csv_headers.dart';
 
 import 'csv_data.dart';
-import 'helpers.dart';
+import '_helpers.dart';
+import '_string_reader.dart';
 
+/// Class used to process CSV data provided a stream of strings. Strings may arrive in variable
+/// chunks and the CSV processor will properly parse each row.
 class CsvProcessor {
   CsvProcessor(Stream<String> stream, String separator, String endOfLine,
       bool hasHeaders)
-      : _input = stream,
-        _separatorRunes = separator.runes.toList(),
+      : _separatorRunes = separator.runes.toList(),
         _endOfLine = endOfLine,
         _endOfLineRunes = endOfLine.runes.toList(),
         _hasHeaders = hasHeaders {
-    _input.listen(_process).onDone(() => _send(close: true));
+    stream.listen(_process, onError: _forwardError, onDone: _close);
   }
 
-  final Stream<String> _input;
   final String _endOfLine;
   final List<int> _endOfLineRunes;
   final List<int> _separatorRunes;
@@ -25,31 +26,34 @@ class CsvProcessor {
   final StringReader _reader = StringReader();
   final StreamController<CsvData> _controller = StreamController();
 
+  /// Stream emitting one [CsvData] per non-empty CSV row.
   Stream<CsvData> get stream => _controller.stream;
 
+  /// List of headers loaded from the CSV file (if the processor was constructed with `hasHeaders` = `true`).
+  Iterable<String> get headers => _headers?.headers ?? const <String>[];
   CsvHeaders? _headers;
-
-  Iterable<String>? get headers => _headers?.headers;
 
   final _data = <String>[];
 
   static final _quote = '"';
   static final _quoteRunes = _quote.runes.toList();
 
+  void _forwardError(dynamic error, StackTrace st) =>
+      _controller.sink.addError(error, st);
+
+  void _close() => _send(close: true);
+
   void _send({bool close = false}) {
     if (close) {
-      if (_state != CsvState.newLine) {
+      if (_state != _CsvState.newLine) {
         _process(_endOfLine);
       }
-      //   print('close: state = $_state, sr = ${dump(_sr._runes)}, sb = ${dump(_sb)}, _data = ${dump(_data)}');
-      // } else {
-      //   print('send: state = $_state, sr = ${dump(_sr._runes)}, sb = ${dump(_sb)}, _data = ${dump(_data)}');
     }
     if (hasData(_data)) {
       if (_hasHeaders && _headers == null) {
-        _headers = CsvHeaders(_data.toList());
+        _headers = CsvHeaders(_data);
       } else {
-        _controller.sink.add(createCsvData(_data.toList(), _headers));
+        _controller.sink.add(CsvDataImpl.create(_data, _headers));
       }
     }
     _data.clear();
@@ -58,39 +62,35 @@ class CsvProcessor {
     }
   }
 
-  CsvState _state = CsvState.newLine;
+  _CsvState _state = _CsvState.newLine;
   final _sb = StringBuffer();
 
   void _process(String data) {
     _reader.append(data);
     var process = true;
     while (process && _reader.length > 0) {
-      // print('state = $_state, sr = ${dump(_sr._runes)}, sb = ${dump(_sb)}, _data = ${dump(_data)}');
       switch (_state) {
-        case CsvState.newLine:
-          _state = CsvState.newValue;
+        case _CsvState.newLine:
+          _state = _CsvState.newValue;
           break;
-        case CsvState.newValue:
+        case _CsvState.newValue:
           process = _processNewValue();
           break;
-        case CsvState.rawValue:
+        case _CsvState.rawValue:
           process = _processValue();
           break;
-        case CsvState.quotedValue:
+        case _CsvState.quotedValue:
           process = _processEscapedValue();
           break;
-        case CsvState.separator:
+        case _CsvState.separator:
           process = _processSeparator();
           break;
-        case CsvState.endOfLine:
+        case _CsvState.endOfLine:
           process = _processEndOfLine();
           break;
         default:
           throw Exception('Invalid state');
       }
-      // if (!process) {
-      //   print('suspend processing: state = $_state, sr = ${dump(_sr._runes)}, sb = ${dump(_sb)}, _data = ${dump(_data)}');
-      // }
     }
   }
 
@@ -100,28 +100,27 @@ class CsvProcessor {
       case Match.partial:
         // need more characters, suspend processing
         return false;
-      case Match.exact:
+      case Match.full:
         // remove opening quote
-        _reader.remove(0, _quoteRunes.length);
-        _state = CsvState.quotedValue;
+        _reader.discard(_quoteRunes.length);
+        _state = _CsvState.quotedValue;
         return true;
       case Match.none:
-        _state = CsvState.rawValue;
+        _state = _CsvState.rawValue;
         return true;
     }
   }
 
   bool _processValue() {
-    var l = 0;
-    bool suspend = false;
+    var l = 0, suspend = false;
     while (l < _reader.length) {
       var match = _reader.isMatchAt(l, _separatorRunes);
       if (match == Match.partial) {
         // need more characters, suspend processing
         suspend = true;
         break;
-      } else if (match == Match.exact) {
-        _state = CsvState.separator;
+      } else if (match == Match.full) {
+        _state = _CsvState.separator;
         break;
       }
 
@@ -130,8 +129,8 @@ class CsvProcessor {
         // need more characters, suspend processing
         suspend = true;
         break;
-      } else if (match == Match.exact) {
-        _state = CsvState.endOfLine;
+      } else if (match == Match.full) {
+        _state = _CsvState.endOfLine;
         break;
       }
 
@@ -141,8 +140,7 @@ class CsvProcessor {
 
     // save this part
     if (l > 0) {
-      _sb.write(_reader.getString(0, l));
-      _reader.remove(0, l);
+      _sb.write(_reader.takeString(l));
     }
     return !suspend;
   }
@@ -157,12 +155,12 @@ class CsvProcessor {
         suspend = true;
         break;
       }
-      if (match == Match.exact) {
+      if (match == Match.full) {
         match = _reader.isMatchAt(l + _quoteRunes.length, _quoteRunes);
-        if (match == Match.exact) {
+        if (match == Match.full) {
           // unescape and continue
-          _sb.write(_reader.getString(0, l + _quoteRunes.length));
-          _reader.remove(0, l + 2 * _quoteRunes.length);
+          _sb.write(_reader.takeString(l + _quoteRunes.length));
+          _reader.discard(_quoteRunes.length);
           l = 0;
           continue;
         } else if (match == Match.partial) {
@@ -172,8 +170,8 @@ class CsvProcessor {
         }
 
         match = _reader.isMatchAt(l + _quoteRunes.length, _separatorRunes);
-        if (match == Match.exact) {
-          _state = CsvState.separator;
+        if (match == Match.full) {
+          _state = _CsvState.separator;
           break;
         } else if (match == Match.partial) {
           // need more characters, suspend processing
@@ -182,8 +180,8 @@ class CsvProcessor {
         }
 
         match = _reader.isMatchAt(l + _quoteRunes.length, _endOfLineRunes);
-        if (match == Match.exact) {
-          _state = CsvState.endOfLine;
+        if (match == Match.full) {
+          _state = _CsvState.endOfLine;
           break;
         } else if (match == Match.partial) {
           // need more characters, suspend processing
@@ -201,12 +199,11 @@ class CsvProcessor {
 
     if (l > 0) {
       // save this part
-      _sb.write(_reader.getString(0, l));
-      _reader.remove(0, l);
+      _sb.write(_reader.takeString(l));
     }
-    if (_state != CsvState.quotedValue) {
+    if (_state != _CsvState.quotedValue) {
       // remove closing quote
-      _reader.remove(0, _quoteRunes.length);
+      _reader.discard(_quoteRunes.length);
     }
     return !suspend;
   }
@@ -214,63 +211,26 @@ class CsvProcessor {
   bool _processSeparator() {
     _data.add(_sb.toString());
     _sb.clear();
-    _reader.remove(0, _separatorRunes.length);
-    _state = CsvState.newValue;
+    _reader.discard(_separatorRunes.length);
+    _state = _CsvState.newValue;
     return true;
   }
 
   bool _processEndOfLine() {
     _data.add(_sb.toString());
     _sb.clear();
-    _reader.remove(0, _endOfLineRunes.length);
+    _reader.discard(_endOfLineRunes.length);
     _send();
-    _state = CsvState.newLine;
+    _state = _CsvState.newLine;
     return true;
   }
 }
 
-enum CsvState {
+enum _CsvState {
   newLine,
   newValue,
   rawValue,
   quotedValue,
   separator,
   endOfLine,
-}
-
-enum Match {
-  none,
-  partial,
-  exact,
-}
-
-class StringReader {
-  final _runes = <int>[];
-
-  int get length => _runes.length;
-
-  String charAt(int idx) => String.fromCharCode(_runes[idx]);
-
-  int operator [](int idx) => _runes[idx];
-
-  Match isMatchAt(int idx, List<int> runes) {
-    final l = length;
-    final rl = runes.length;
-    for (var ridx = 0; ridx < rl; ridx++) {
-      if (idx >= l) {
-        return Match.partial;
-      } else if (_runes[idx] != runes[ridx]) {
-        return Match.none;
-      }
-      idx++;
-    }
-    return Match.exact;
-  }
-
-  void remove(int start, int end) => _runes.removeRange(start, end);
-
-  String getString(int start, int end) =>
-      String.fromCharCodes(_runes, start, end);
-
-  void append(String text) => _runes.addAll(text.runes);
 }