+++ lib/src/csv_data.dart
@@ -1,13 +1,15 @@
 import 'dart:math' as math show max;
+import 'package:meta/meta.dart';
 
-import 'csv_headers.dart';
+import '_csv_headers.dart';
+import 'csv_loader.dart';
 import 'exceptions.dart';
-import 'helpers.dart';
+import '_helpers.dart';
 
-/// [CsvData] holds the values from a CSV line. If the [CsvLoader] was created with [CsvLoader.withHeaders], it
-/// also holds a reference to the list of headers found in the CSV's first non-empty line. Values can be accessed
-/// by index (absolute value), or (when created with [CsvLoader.withHeaders]) by header with relative index when
-/// necessary (i.e. when multiple headers have the same label).
+/// [CsvData] holds the values from a CSV line. If the [CsvLoader] was created with [CsvLoader.withHeaders],
+/// it also holds a reference to the list of headers found in the CSV's first non-empty row. Values can
+/// be accessed by index (absolute value), or (when created with [CsvLoader.withHeaders]) by header with
+/// relative index when necessary (i.e. when several headers have the same label).
 ///
 /// Example:
 /// ```dart
@@ -41,7 +43,8 @@ import 'helpers.dart';
 /// }
 /// ```
 class CsvData {
-  CsvData._(this._values, [this._headers]);
+  CsvData._(List<String> values, [this._headers])
+      : _values = List.unmodifiable(values);
 
   final CsvHeaders? _headers;
   final List<String> _values;
@@ -55,13 +58,13 @@ class CsvData {
   int get columnCount => _headers?.count ?? 0;
 
   /// List of values. Note that `values.length` may be different from [columnCount], e.g. if the reader
-  /// was not constructed with [CsvLoader.withHeaders], or the CSV line has more or less values than
+  /// was not constructed with [CsvLoader.withHeaders], or the CSV row has more or less values than
   /// headers.
   Iterable<String> get values => _values;
 
   int _getValueIndex([String header = '', int index = -1]) {
-    if (isNullOrEmpty(header)) {
-      var max = math.max(_headers?.count ?? 0, _values.length);
+    if (isEmptyOrWhiteSpace(header)) {
+      final max = math.max(_headers?.count ?? 0, _values.length);
       if (index < 0 || index >= max) {
         throw InvalidHeaderException(
             'Header "$index" out of range (0..${max - 1})');
@@ -76,7 +79,7 @@ class CsvData {
 
   /// Gets value for [header] from the current record. If [header] is an [int], it is interpreted as the
   /// column index. If [header] is a [String], it is used to lookup the header and find the column index.
-  dynamic operator [](dynamic header) {
+  String? operator [](dynamic header) {
     if (header is int) {
       return get('', header);
     } else if (header is String) {
@@ -87,13 +90,14 @@ class CsvData {
     }
   }
 
-  /// Gets value for [header] / [index] from the current record. The column index in the CSV record is retrieved
-  /// according to [header] and [index]. If [header] is not set, [index] is used as the column index (starting from
-  /// 0). If [header] is set, the column index will be retrieved from headers, provided the reader was constructed
-  /// with [CsvLoader.withHeaders]. If multiple headers have the same label, [index] can be used to distinguish
-  /// amongst them (starting from 0). If no match is found, or if [index] is out of bounds, throws an
-  /// [InvalidHeaderException]. Returns `null` if the CSV line has fewer values than headers, returns `null`.
-  dynamic get([String header = '', int index = -1]) {
+  /// Gets value for [header] / [index] from the current record. The column index in the CSV record is
+  /// retrieved according to [header] and [index]. If [header] is not set, [index] is used as the column
+  /// index (starting from 0). If [header] is set, the column index will be retrieved from headers,
+  /// provided the reader was constructed with [CsvLoader.withHeaders]. If several headers have the
+  /// same label, [index] can be used to distinguish amongst them (starting from 0). If no match is
+  /// found, or if [index] is out of bounds, throws an [InvalidHeaderException]. Returns `null` if the
+  /// CSV line has fewer values than headers.
+  String? get([String header = '', int index = -1]) {
     final idx = _getValueIndex(header, index);
     if (0 <= idx && idx < _values.length) {
       return _values[idx];
@@ -104,5 +108,8 @@ class CsvData {
 }
 
 // for internal use, do not export
-CsvData createCsvData(List<String> values, [CsvHeaders? headers]) =>
-    CsvData._(values, headers);
+@internal
+extension CsvDataImpl on CsvData {
+  static CsvData create(List<String> values, [CsvHeaders? headers]) =>
+      CsvData._(values, headers);
+}